Finished building simon says mode and finished building a crude terminal based client for the remote mode. This needs to be put into an app

+++ Pi/src/games/Game.java
@@ -3,5 +3,5 @@ package games;
 import java.util.Observer;
 
 public interface Game extends Observer {
-	public int playGame();
+	public long playGame();
 }
